requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
pydantic
python-dotenv
ipython
pytest
httpx
pytest-asyncio
asyncio
psycopg2==2.9.10